Debug: POST /players/{id}/games actualitza player winsPercent

+++ app/models/player.js
@@ -10,7 +10,7 @@ Player.init({
     allowNull: false
   },
   winsPercent: {
-    type: DataTypes.FLOAT,
+    type: DataTypes.DECIMAL(5,2), // Cal que sigui decimal per poder filtrar a SQL perquè, si li posem FLOAT, clàusula WHERE retorna null quan es compara per igualtat encara que el valor existeixi (per la imprecisió dels floats)
     defaultValue: null,  // Faig servir null com a 'no inicialitzat'
     allowNull: true
   }
